requirements-dev.txt
setupext-janitor
ipython
pydb
